all references to 'count' changed to 'NSUInteger'
git-svn-id: svn://orca.physics.unc.edu/Orca64@11885 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Hardware/USB/LakeShore336/ORLakeShore336Heater.h
@@ -51,7 +51,7 @@
 - (id)   initWithCoder:(NSCoder*)decoder;
 - (void) encodeWithCoder:(NSCoder*)encoder;
 
-- (int)  numberPointsInTimeRate;
+- (NSUInteger)  numberPointsInTimeRate;
 - (void) timeRateAtIndex:(int)i x:(double*)xValue y:(double*)yValue;
 - (NSString*) heaterSetupString;
 - (NSString*) pidSetupString;
